requirements.txt
fastapi>=0.111.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
httpx>=0.25.0,<0.28.0
websockets>=11.0,<13.0
pydantic>=2.6,<3.0
orjson>=3.9,<4.0
prometheus-client>=0.16,<0.21
tenacity>=8.2,<9.0
aiolimiter>=1.1,<2.0
PyYAML>=6.0,<7.0
typing_extensions>=4.8,<5.0

# Dev and testing
pytest>=7.4,<9.0
pytest-asyncio>=0.21,<0.24
